Fixed smarthome id beeing uint instead of long for get config
* because at some point all ids have been exchanged for longs, because the IEEE id of smarthome devices most of times are longs

+++ AppBrokerASP/Smarthome.cs
@@ -58,7 +58,7 @@ public class SmartHome : Hub<ISmartHomeClient>
 
     }
 
-    public dynamic? GetConfig(uint deviceId) => IInstanceContainer.Instance.DeviceManager.Devices.TryGetValue(deviceId, out Device? device) ? device.GetConfig() : null;
+    public dynamic? GetConfig(long deviceId) => IInstanceContainer.Instance.DeviceManager.Devices.TryGetValue(deviceId, out Device? device) ? device.GetConfig() : null;
 
     public async void SendUpdate(Device device) => await (Clients.All?.Update(device) ?? Task.CompletedTask);
